+++ ot/gromov/__init__.py
@@ -47,7 +47,7 @@ from ._dictionary import (gromov_wasserstein_dictionary_learning,
                           fused_gromov_wasserstein_dictionary_learning,
                           fused_gromov_wasserstein_linear_unmixing)
 
-from ._lowrank import (_flat_product_operator, lowrank_gromov_wasserstein)
+from ._lowrank import (_flat_product_operator, lowrank_gromov_wasserstein_samples)
 
 
 __all__ = ['init_matrix', 'tensor_product', 'gwloss', 'gwggrad', 'update_square_loss',
@@ -66,4 +66,4 @@ __all__ = ['init_matrix', 'tensor_product', 'gwloss', 'gwggrad', 'update_square_
            'entropic_semirelaxed_gromov_wasserstein2', 'entropic_semirelaxed_fused_gromov_wasserstein',
            'entropic_semirelaxed_fused_gromov_wasserstein2', 'gromov_wasserstein_dictionary_learning',
            'gromov_wasserstein_linear_unmixing', 'fused_gromov_wasserstein_dictionary_learning',
-           'fused_gromov_wasserstein_linear_unmixing']
+           'fused_gromov_wasserstein_linear_unmixing', 'lowrank_gromov_wasserstein_samples']